requirements_updated.txt
Flask==3.0.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0